Upgrade to 11.0

+++ src/Avalonia.Samples/MVVM/ValueConversionSample/Converter/MathAddConverter.cs
@@ -13,13 +13,13 @@ namespace ValueConversionSample.Converter
         {
             // For add this is simple. just return the sum of the value and the parameter.
             // You may want to validate value and parameter in a real world App
-            return (double?)value + (double?)parameter;
+            return (decimal?)value + (decimal?)parameter;
         }
 
         public object? ConvertBack(object? value, Type targetType, object? parameter, CultureInfo culture)
         {
             // If we want to convert back, we need to subtract instead of add.
-            return (double?)value - (double?)parameter;
+            return (decimal?)value - (decimal?)parameter;
         }
     }
 }
